fix code after code-review
